requirements.txt
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
filelock==3.18.0
flake8==7.1.2
identify==2.6.9
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.29.3
